pyproject.toml
[project]
name = "text-image-rag"
version = "0.1.0"
description = "Industry-grade multimodal RAG system for text, tables, and images"
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    # --- Frontend ---
    "chainlit>=2.0.0",

    # --- API Backend ---
    "fastapi>=0.115.0",
    "uvicorn>=0.25.0",
    "pydantic-settings>=2.3.0",
    "slowapi>=0.1.9",
    "python-jose[cryptography]>=3.3.0",

    # --- Document Parsing ---
    "pymupdf>=1.27.2",
    "pdfplumber>=0.11.0",
    "python-docx>=1.2.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",

    # --- Embedding & Retrieval ---
    "sentence-transformers>=3.0.0",
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "qdrant-client>=1.9.0",
    "rank-bm25>=0.2.2",
    "langdetect>=1.0.9",

    # --- LLM & Orchestration ---
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-groq>=0.2.0",
    "groq>=0.9.0",
    "langsmith>=0.1.0",
    "tiktoken>=0.7.0",

    # --- Async Processing & Caching ---
    "celery>=5.4.0",
    "redis>=5.0.0",

    # --- Evaluation ---
    "ragas>=0.1.0",

    # --- Logging & Config ---
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]
